+++ psiz/simulate.py
@@ -67,7 +67,7 @@ class Agent(object):
                 order of the stimuli is now informative.
 
         """
-        group_id = self.group_id * np.ones((trials.n_trial), dtype=np.int)
+        group_id = self.group_id * np.ones((trials.n_trial), dtype=np.int32)
         prob_all = self.embedding.outcome_probability(
             trials, group_id=group_id)
         judged_trials = self._select(trials, prob_all)
@@ -93,9 +93,9 @@ class Agent(object):
         n_config = trials.config_list.shape[0]
 
         # Pre-allocate.
-        chosen_outcome_idx = np.empty((n_trial_all), dtype=np.int64)
+        chosen_outcome_idx = np.empty((n_trial_all), dtype=np.int32)
         stimulus_set = -1 * np.ones(
-            (n_trial_all, 1 + max_n_ref), dtype=np.int64
+            (n_trial_all, 1 + max_n_ref), dtype=np.int32
         )
         stimulus_set[:, 0] = trials.stimulus_set[:, 0]
         for i_config in range(n_config):
@@ -115,7 +115,7 @@ class Agent(object):
                 stimulus_set[trial_idx[i_trial], 1:n_reference+1] = \
                     stimuli_set_ref[i_trial, outcome_idx[outcome_loc, :]]
 
-        group_id = np.full((trials.n_trial), self.group_id, dtype=np.int64)
+        group_id = np.full((trials.n_trial), self.group_id, dtype=np.int32)
         return JudgedTrials(
                 stimulus_set,
                 n_selected=trials.n_selected,
